🔵 move function to "helpers/objects.ts"
